requirements.txt
streamlit
pydantic>=2.0
pandas
pdfminer.six
anthropic
requests
beautifulsoup4
